feat(peer msgs): individual send rates

+++ sn/src/messaging/system/mod.rs
@@ -87,7 +87,7 @@ pub enum SystemMsg {
     AntiEntropyProbe(XorName),
     /// Sent when a msg-consuming node wants to update a msg-producing node on the number of msgs per s it wants to receive.
     /// It tells the node to adjust msg sending rate according to the provided value in this msg.
-    BackPressure(usize),
+    BackPressure(f64),
     /// Send from a section to the node to be immediately relocated.
     Relocate(SectionAuth<NodeState>),
     /// Sent from a bootstrapping peer to the section requesting to join as a new member
